+++ src/main/java/events/UserJoinEventListener.java
@@ -18,11 +18,11 @@ import java.time.ZonedDateTime;
 public class UserJoinEventListener {
     @EventSubscriber
     public void onUserJoinEvent(UserJoinEvent event) throws RateLimitException, DiscordException, MissingPermissionsException, IOException {
-        if (FileManager.checkIfAlreadyExists(event.getGuild()) != null) {
-            if (ZonedDateTime.now(ZoneId.of(FileManager.checkIfAlreadyExists(event.getGuild()))).getMinute() >= 10)
-                MessageSender.sendMessage("User " + event.getUser().mention() + " just joined " + event.getGuild().getName() + "!" + " ```" + ZonedDateTime.now(ZoneId.of(FileManager.checkIfAlreadyExists(event.getGuild()))).getDayOfMonth() + "." + ZonedDateTime.now(ZoneId.of(FileManager.checkIfAlreadyExists(event.getGuild()))).getMonthValue() + "." + ZonedDateTime.now(ZoneId.of(FileManager.checkIfAlreadyExists(event.getGuild()))).getYear() + " " + ZonedDateTime.now(ZoneId.of(FileManager.checkIfAlreadyExists(event.getGuild()))).getHour() + ":" + ZonedDateTime.now(ZoneId.of(FileManager.checkIfAlreadyExists(event.getGuild()))).getMinute() + "```", event.getGuild().getChannelByID(event.getGuild().getID()));
+        if (FileManager.searchInFile("timeZoneDB", event.getGuild().getID()) != null) {
+            if (ZonedDateTime.now(ZoneId.of(FileManager.searchInFile("timeZoneDB", event.getGuild().getID()))).getMinute() >= 10)
+                MessageSender.sendMessage("User " + event.getUser().mention() + " just joined " + event.getGuild().getName() + "!" + " ```" + ZonedDateTime.now(ZoneId.of(FileManager.searchInFile("timeZoneDB", event.getGuild().getID()))).getDayOfMonth() + "." + ZonedDateTime.now(ZoneId.of(FileManager.searchInFile("timeZoneDB", event.getGuild().getID()))).getMonthValue() + "." + ZonedDateTime.now(ZoneId.of(FileManager.searchInFile("timeZoneDB", event.getGuild().getID()))).getYear() + " " + ZonedDateTime.now(ZoneId.of(FileManager.searchInFile("timeZoneDB", event.getGuild().getID()))).getHour() + ":" + ZonedDateTime.now(ZoneId.of(FileManager.searchInFile("timeZoneDB", event.getGuild().getID()))).getMinute() + "```", event.getGuild().getChannelByID(event.getGuild().getID()));
             else
-                MessageSender.sendMessage("User " + event.getUser().mention() + " just joined " + event.getGuild().getName() + "!" + " ```" + ZonedDateTime.now(ZoneId.of(FileManager.checkIfAlreadyExists(event.getGuild()))).getDayOfMonth() + "." + ZonedDateTime.now(ZoneId.of(FileManager.checkIfAlreadyExists(event.getGuild()))).getMonthValue() + "." + ZonedDateTime.now(ZoneId.of(FileManager.checkIfAlreadyExists(event.getGuild()))).getYear() + " " + ZonedDateTime.now(ZoneId.of(FileManager.checkIfAlreadyExists(event.getGuild()))).getHour() + ":0" + ZonedDateTime.now(ZoneId.of(FileManager.checkIfAlreadyExists(event.getGuild()))).getMinute() + "```", event.getGuild().getChannelByID(event.getGuild().getID()));
+                MessageSender.sendMessage("User " + event.getUser().mention() + " just joined " + event.getGuild().getName() + "!" + " ```" + ZonedDateTime.now(ZoneId.of(FileManager.searchInFile("timeZoneDB", event.getGuild().getID()))).getDayOfMonth() + "." + ZonedDateTime.now(ZoneId.of(FileManager.searchInFile("timeZoneDB", event.getGuild().getID()))).getMonthValue() + "." + ZonedDateTime.now(ZoneId.of(FileManager.searchInFile("timeZoneDB", event.getGuild().getID()))).getYear() + " " + ZonedDateTime.now(ZoneId.of(FileManager.searchInFile("timeZoneDB", event.getGuild().getID()))).getHour() + ":0" + ZonedDateTime.now(ZoneId.of(FileManager.searchInFile("timeZoneDB", event.getGuild().getID()))).getMinute() + "```", event.getGuild().getChannelByID(event.getGuild().getID()));
         }else{
             MessageSender.sendMessage("```Phost\nCouldn't send message with join info because no timezone was set on this server! To do so type '♥settimezone <yourTimezone>'\n```", event.getGuild().getChannelByID(event.getGuild().getID()));
         }
