+++ src/main/java/events/UserLeaveEventListener.java
@@ -20,11 +20,11 @@ public class UserLeaveEventListener {
 
     @EventSubscriber
     public void onUserLeaveEvent(UserLeaveEvent event) throws IOException, RateLimitException, DiscordException, MissingPermissionsException {
-        if (FileManager.checkIfAlreadyExists(event.getGuild()) != null) {
-            if (ZonedDateTime.now(ZoneId.of(FileManager.checkIfAlreadyExists(event.getGuild()))).getMinute() >= 10)
-                MessageSender.sendMessage("User " + event.getUser().mention() + " just left/got kicked from " + event.getGuild().getName() + "!" + " ```" + ZonedDateTime.now(ZoneId.of(FileManager.checkIfAlreadyExists(event.getGuild()))).getDayOfMonth() + "." + ZonedDateTime.now(ZoneId.of(FileManager.checkIfAlreadyExists(event.getGuild()))).getMonthValue() + "." + ZonedDateTime.now(ZoneId.of(FileManager.checkIfAlreadyExists(event.getGuild()))).getYear() + " " + ZonedDateTime.now(ZoneId.of(FileManager.checkIfAlreadyExists(event.getGuild()))).getHour() + ":" + ZonedDateTime.now(ZoneId.of(FileManager.checkIfAlreadyExists(event.getGuild()))).getMinute() + "```", event.getGuild().getChannelByID(event.getGuild().getID()));
+        if (FileManager.searchInFile("timeZoneDB", event.getGuild().getID()) != null) {
+            if (ZonedDateTime.now(ZoneId.of(FileManager.searchInFile("timeZoneDB", event.getGuild().getID()))).getMinute() >= 10)
+                MessageSender.sendMessage("User " + event.getUser().mention() + " just left/got kicked from " + event.getGuild().getName() + "!" + " ```" + ZonedDateTime.now(ZoneId.of(FileManager.searchInFile("timeZoneDB", event.getGuild().getID()))).getDayOfMonth() + "." + ZonedDateTime.now(ZoneId.of(FileManager.searchInFile("timeZoneDB", event.getGuild().getID()))).getMonthValue() + "." + ZonedDateTime.now(ZoneId.of(FileManager.searchInFile("timeZoneDB", event.getGuild().getID()))).getYear() + " " + ZonedDateTime.now(ZoneId.of(FileManager.searchInFile("timeZoneDB", event.getGuild().getID()))).getHour() + ":" + ZonedDateTime.now(ZoneId.of(FileManager.searchInFile("timeZoneDB", event.getGuild().getID()))).getMinute() + "```", event.getGuild().getChannelByID(event.getGuild().getID()));
             else
-                MessageSender.sendMessage("User " + event.getUser().mention() + " just left/got kicked from " + event.getGuild().getName() + "!" + " ```" + ZonedDateTime.now(ZoneId.of(FileManager.checkIfAlreadyExists(event.getGuild()))).getDayOfMonth() + "." + ZonedDateTime.now(ZoneId.of(FileManager.checkIfAlreadyExists(event.getGuild()))).getMonthValue() + "." + ZonedDateTime.now(ZoneId.of(FileManager.checkIfAlreadyExists(event.getGuild()))).getYear() + " " + ZonedDateTime.now(ZoneId.of(FileManager.checkIfAlreadyExists(event.getGuild()))).getHour() + ":0" + ZonedDateTime.now(ZoneId.of(FileManager.checkIfAlreadyExists(event.getGuild()))).getMinute() + "```", event.getGuild().getChannelByID(event.getGuild().getID()));
+                MessageSender.sendMessage("User " + event.getUser().mention() + " just left/got kicked from " + event.getGuild().getName() + "!" + " ```" + ZonedDateTime.now(ZoneId.of(FileManager.searchInFile("timeZoneDB", event.getGuild().getID()))).getDayOfMonth() + "." + ZonedDateTime.now(ZoneId.of(FileManager.searchInFile("timeZoneDB", event.getGuild().getID()))).getMonthValue() + "." + ZonedDateTime.now(ZoneId.of(FileManager.searchInFile("timeZoneDB", event.getGuild().getID()))).getYear() + " " + ZonedDateTime.now(ZoneId.of(FileManager.searchInFile("timeZoneDB", event.getGuild().getID()))).getHour() + ":0" + ZonedDateTime.now(ZoneId.of(FileManager.searchInFile("timeZoneDB", event.getGuild().getID()))).getMinute() + "```", event.getGuild().getChannelByID(event.getGuild().getID()));
         }else{
             MessageSender.sendMessage("```Phost\nCouldn't send message with leave/kick info because no timezone was set on this server! To do so type '♥settimezone <yourTimezone>'\n```", event.getGuild().getChannelByID(event.getGuild().getID()));
         }
