+++ src/main/java/main.java
@@ -1,8 +1,5 @@
 import events.*;
 import client.ClientManager;
-import console.ConsoleCommandManager;
-import utils.FileManager;
-import utils.FileManager2;
 
 import java.io.IOException;
 import java.util.Scanner;
@@ -19,7 +16,6 @@ public class main {
 
         ClientManager.setDeveloperMode(true);   //THIS TELLS THE BOT WHETHER HE IS IN DEV MODE OR NOT
         System.out.println("Please enter your bot Token");
-        FileManager.createFileObject();
         ClientManager.newClientInstance(new Scanner(System.in).next(), true);
         ClientManager.newDispatcher();
         ClientManager.getDispatcher().registerListener(new ReadyEventListener());
@@ -27,7 +23,5 @@ public class main {
         ClientManager.getDispatcher().registerListener(new UserJoinEventListener());
         ClientManager.getDispatcher().registerListener(new UserBanEventListener());
         ClientManager.getDispatcher().registerListener(new UserLeaveEventListener());
-        FileManager2.writeIntoFile(FileManager2.getAndCreateFile("Dieter"), "MoinAPO", "TheseNUTZZZZZ");
-        System.out.println(FileManager2.searchInFile(FileManager2.getAndCreateFile("Dieter"), "MoinAPO"));
     }
 }
